update code cmm013

+++ nts.uk/web/nts.uk.com.web/src/main/webapp/view/cmm/013/a/cmm013.a.vm.js
@@ -10,32 +10,21 @@ var cmm013;
                     self.jTitleRef = ko.observableArray([]);
                     self.dataRef = ko.observableArray([]);
                     self.dataRefNew = ko.observableArray([]);
-                    self.label_002 = ko.observable(new Labels());
                     self.label_003 = ko.observable(new Labels());
-                    self.label_004 = ko.observable(new Labels());
-                    self.label_005 = ko.observable(new Labels());
-                    self.label_006 = ko.observable(new Labels());
-                    self.sel_002 = ko.observable('');
                     self.inp_002_code = ko.observable(null);
                     self.inp_002_enable = ko.observable(null);
                     self.inp_003_name = ko.observable(null);
                     self.inp_005_memo = ko.observable(null);
                     self.createdMode = ko.observable(true);
-                    self.updatedata = ko.observable(null);
-                    self.adddata = ko.observable(null);
                     self.listbox = ko.observableArray([]);
                     self.selectedCode = ko.observable('');
                     self.isEnable = ko.observable(true);
                     self.itemHist = ko.observable(null);
-                    self.itemName = ko.observable('');
                     self.index_selected = ko.observable('');
                     self.srtDateLast = ko.observable(null);
                     self.startDateLast = ko.observable(null);
-                    self.historyIdLast = ko.observable(null);
-                    self.length = ko.observable(0);
                     self.startDateAddNew = ko.observable("");
                     self.checkRegister = ko.observable('0');
-                    self.startDateUpdate = ko.observable(null);
                     self.oldStartDate = ko.observable(null);
                     self.endDateUpdate = ko.observable(null);
                     self.histIdUpdate = ko.observable(null);
@@ -43,16 +32,11 @@ var cmm013;
                     self.oldEndDate = ko.observable(null);
                     self.startDateLastEnd = ko.observable(null);
                     self.checkCoppyJtitle = ko.observable(true);
-                    self.checkAddJhist = ko.observable('0');
-                    self.checkAddJtitle = ko.observable('0');
-                    self.checkUpdate = ko.observable('0');
-                    self.checkDelete = ko.observable('0');
                     self.historyIdUpdate = ko.observable(null);
                     self.dataSource = ko.observableArray([]);
                     self.currentItem = ko.observable(null);
-                    self.itemName = ko.observable('');
                     self.currentCode = ko.observable();
-                    self.changeDetail = ko.observable(false);
+                    self.clickChange = ko.observable(false);
                     self.columns = ko.observableArray([
                         { headerText: 'コード', key: 'jobCode', width: 80 },
                         { headerText: '名称', key: 'jobName', width: 100 }
@@ -69,6 +53,11 @@ var cmm013;
                     ]);
                     self.selectedId = ko.observable(0);
                     self.enable = ko.observable(true);
+                    self.notAlert = ko.observable(true);
+                    self.dirty = new nts.uk.ui.DirtyChecker(self.listbox);
+                    $("#list-box").click(function (evt, ui) {
+                        self.clickChange(true);
+                    });
                     self.selectedCode.subscribe(function (codeChanged) {
                         self.itemHist(self.findHist(codeChanged));
                         self.oldStartDate(self.itemHist().startDate);
@@ -89,70 +78,84 @@ var cmm013;
                             a.service.findAllPosition(codeChanged).done(function (position_arr) {
                                 self.dataSource(position_arr);
                                 if (self.dataSource().length > 0) {
-                                    self.currentCode(self.inp_002_code() || self.dataSource()[0].jobCode);
-                                    self.changeDetail(true);
+                                    var changedCode = self.clickChange() ? self.dataSource()[0].jobCode : self.inp_002_code() || self.dataSource()[0].jobCode;
+                                    if (changedCode === self.currentCode()) {
+                                        self.changedCode(changedCode);
+                                    }
+                                    else {
+                                        self.currentCode(changedCode);
+                                    }
                                 }
-                                else {
-                                    self.changeDetail(false);
-                                }
+                                self.clickChange(false);
                             }).fail(function (err) {
                                 nts.uk.ui.dialog.alert(err.message);
                             });
                         }
                     });
                     self.currentCode.subscribe(function (codeChanged) {
-                        self.changeDetail(true);
-                    });
-                    self.changeDetail.subscribe(function (value) {
-                        if (value) {
-                            self.currentItem(self.findPosition(self.currentCode()));
-                            if (self.currentItem() != null) {
-                                self.inp_002_code(self.currentItem().jobCode);
-                                self.inp_003_name(self.currentItem().jobName);
-                                self.inp_005_memo(self.currentItem().memo);
-                                self.selectedId(self.currentItem().presenceCheckScopeSet);
-                                self.inp_002_enable(false);
-                                self.createdMode(false);
-                            }
-                            a.service.getAllJobTitleAuth(self.currentItem().historyId, self.currentItem().jobCode).done(function (jTref) {
-                                if (jTref.length === 0) {
-                                    $('.trLst003').css('visibility', 'hidden');
-                                }
-                                else {
-                                    $('.trLst003').css('visibility', 'visible');
-                                    self.dataRef([]);
-                                    _.map(jTref, function (item) {
-                                        var tmp = new model.GetAuth(item.jobCode, item.authCode, item.authName, item.referenceSettings);
-                                        self.dataRef.push(tmp);
-                                        return tmp;
-                                    });
-                                    self.dataRefNew(self.dataRef());
-                                    self.createdMode(true);
-                                    self.changeDetail(false);
-                                }
-                            });
+                        if (codeChanged !== null && codeChanged !== undefined) {
+                            self.changedCode(codeChanged);
                         }
                     });
                 }
+                ScreenModel.prototype.changedCode = function (value) {
+                    var self = this;
+                    self.currentItem(self.findPosition(value));
+                    if (self.currentItem() != null) {
+                        self.inp_002_code(self.currentItem().jobCode);
+                        self.inp_003_name(self.currentItem().jobName);
+                        self.inp_005_memo(self.currentItem().memo);
+                        self.selectedId(self.currentItem().presenceCheckScopeSet);
+                        self.inp_002_enable(false);
+                        self.createdMode(false);
+                        a.service.getAllJobTitleAuth(self.currentItem().historyId, self.currentItem().jobCode).done(function (jTref) {
+                            if (jTref.length === 0) {
+                                $('.trLst003').hide();
+                            }
+                            else {
+                                $('.trLst003').show();
+                                self.dataRef([]);
+                                _.map(jTref, function (item) {
+                                    var tmp = new model.GetAuth(item.jobCode, item.authCode, item.authName, item.referenceSettings);
+                                    self.dataRef.push(tmp);
+                                    return tmp;
+                                });
+                                self.dataRefNew(self.dataRef());
+                                self.createdMode(true);
+                            }
+                        });
+                    }
+                };
                 ScreenModel.prototype.startPage = function () {
                     var self = this;
                     var dfd = $.Deferred();
                     self.getHistory(dfd);
                     return dfd.promise();
                 };
-                ScreenModel.prototype.getHistory = function (dfd) {
+                ScreenModel.prototype.getHistory = function (dfd, selectedHistory) {
                     var self = this;
                     a.service.getAllHistory().done(function (history_arr) {
                         if (history_arr.length > 0) {
                             self.listbox(history_arr);
-                            var histStart = _.first(history_arr);
+                            if (selectedHistory !== undefined && selectedHistory !== "1") {
+                                var currentHist = self.findHist(selectedHistory);
+                                self.selectedCode(currentHist.historyId);
+                                self.srtDateLast(currentHist.startDate);
+                                self.endDateUpdate(currentHist.endDate);
+                                self.histIdUpdate(currentHist.historyId);
+                                self.startDateLast(currentHist.startDate);
+                                self.historyIdUpdate(currentHist.historyId);
+                            }
+                            else {
+                                var histStart = _.first(history_arr);
+                                self.selectedCode(histStart.historyId);
+                                self.srtDateLast(histStart.startDate);
+                                self.endDateUpdate(histStart.endDate);
+                                self.histIdUpdate(histStart.historyId);
+                                self.startDateLast(histStart.startDate);
+                                self.historyIdUpdate(histStart.historyId);
+                            }
                             var hisEnd = _.last(history_arr);
-                            self.selectedCode(histStart.historyId);
-                            self.srtDateLast(histStart.startDate);
-                            self.endDateUpdate(histStart.endDate);
-                            self.histIdUpdate(histStart.historyId);
-                            self.startDateLast(histStart.startDate);
-                            self.historyIdUpdate(histStart.historyId);
                             self.startDateLastEnd(hisEnd.startDate);
                             self.oldStartDate();
                             dfd.resolve(history_arr);
@@ -194,12 +197,13 @@ var cmm013;
                         positionInfor.jobCode = self.inp_002_code();
                         positionInfor.chkInsert = self.inp_002_enable();
                         positionInfor.positionCommand = jobInfor;
+                        var selectedHistory_1 = self.selectedCode();
                         a.service.registry(positionInfor).done(function () {
                             nts.uk.ui.windows.setShared('cmm013Insert', '', true);
                             nts.uk.ui.windows.setShared('cmm013Copy', '', true);
                             nts.uk.ui.windows.setShared('cmm013C_startDateNew', '', true);
                             self.selectedCode.valueHasMutated();
-                            self.getHistory(dfd);
+                            self.getHistory(dfd, selectedHistory_1);
                         });
                         return dfd.promise();
                     }
@@ -234,16 +238,32 @@ var cmm013;
                 ScreenModel.prototype.initPosition = function () {
                     var self = this;
                     if (self.checkChangeData() == false || self.checkChangeData() === undefined) {
-                        self.inp_002_enable(true);
-                        self.inp_002_code("");
-                        self.inp_003_name("");
-                        self.selectedId(1);
-                        self.inp_005_memo("");
-                        self.currentCode(null);
+                        if (self.dirty.isDirty()) {
+                            nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function () {
+                                self.inp_002_enable(true);
+                                self.inp_002_code("");
+                                self.inp_003_name("");
+                                self.selectedId(1);
+                                self.inp_005_memo("");
+                                self.currentCode(null);
+                                $("#inp_002").focus();
+                            }).ifNo(function () {
+                            });
+                        }
+                        else {
+                            self.inp_002_enable(true);
+                            self.inp_002_code("");
+                            self.inp_003_name("");
+                            self.selectedId(1);
+                            self.inp_005_memo("");
+                            self.currentCode(null);
+                            $("#inp_002").focus();
+                        }
                     }
                 };
                 ScreenModel.prototype.checkChangeData = function () {
                     var self = this;
+                    var dfd = $.Deferred();
                     if (self.checkRegister() == '1') {
                         var retVal = confirm("Changed ?");
                         if (retVal == true && self.startDateAddNew() !== undefined && self.startDateAddNew() != '') {
@@ -251,36 +271,13 @@ var cmm013;
                             return true;
                         }
                         else {
+                            var selectedHistory = self.selectedCode();
                             self.startDateAddNew('');
                             self.checkRegister('0');
-                            self.getAllJobHistAfterHandler();
+                            self.getHistory(dfd, selectedHistory);
                             return false;
                         }
                     }
-                };
-                ScreenModel.prototype.getAllJobHistAfterHandler = function () {
-                    var self = this;
-                    var dfd = $.Deferred();
-                    a.service.getAllHistory().done(function (history_arr) {
-                        if (history_arr.length > 0) {
-                            self.listbox(history_arr);
-                            var histStart = _.first(history_arr);
-                            var hisEnd = _.last(history_arr);
-                            self.selectedCode(histStart.startDate);
-                            self.srtDateLast(histStart.startDate);
-                            self.endDateUpdate(histStart.endDate);
-                            self.histIdUpdate(histStart.historyId);
-                            self.startDateLast(histStart.startDate);
-                            self.historyIdUpdate(histStart.historyId);
-                            self.startDateLastEnd(hisEnd.startDate);
-                            dfd.resolve(history_arr);
-                        }
-                        else {
-                            self.openCDialog();
-                            dfd.resolve();
-                        }
-                    });
-                    return dfd.promise();
                 };
                 ScreenModel.prototype.openCDialog = function () {
                     var self = this;
@@ -373,10 +370,14 @@ var cmm013;
                         var dfd = $.Deferred();
                         var item = new model.DeleteJobTitle(self.selectedCode(), self.currentItem().jobCode);
                         self.index_of_itemDelete = self.dataSource().indexOf(self.currentItem());
-                        a.service.deletePosition(item).done(function (res) {
-                            self.getPositionList_afterDelete();
-                        }).fail(function (res) {
-                            dfd.reject(res);
+                        nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function () {
+                            a.service.deletePosition(item).done(function (res) {
+                                self.dirty.reset();
+                                self.getPositionList_afterDelete();
+                            }).fail(function (res) {
+                                dfd.reject(res);
+                            });
+                        }).ifNo(function () {
                         });
                     }
                 };
